pyproject.toml
[tool.poetry]
name = "angel_system"
version = "0.1.0"
description = ""
authors = ["Kitware, Inc. <kitware@kitware.com>"]

[tool.poetry.dependencies]
python = "^3.8,<3.10"
matplotlib = "^3.5.1"
numpy = "^1.22.2"
Pillow = "^8"
simpleaudio = "^1.0.4"
smqtk-detection = ">=0.19.0"
torch = "^1.10.2"
torchvision = "^0.11.3"
trimesh = {extras = ["all"], version = "^3.10.2"}
pytorchvideo = "^0.1.5"
transitions = "^0.8.11"
timm = "^0.5.4"
einops = "^0.4.1"
pynput = "^1.7.6"
scriptconfig = "^0.6.4"
pandas = "^1.4.3"
pyarrow = "^9.0.0"

[tool.poetry.dev-dependencies]
ipython = "^8.0.1"

[build-system]
requires = ["poetry-core==1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ptg_eval = "angel_system.eval.ptg_evaluate:main"

[tool.poetry.plugins."smqtk_plugins"]
# Activity detector
"angel_system.impls.detect_activities.pytorchvideo_slow_fast_r50" = "angel_system.impls.detect_activities.pytorchvideo_slow_fast_r50"
"angel_system.impls.detect_activities.swinb.swinb_detect_activities" = "angel_system.impls.detect_activities.swinb.swinb_detect_activities"
"angel_system.impls.detect_activities.two_stage.two_stage_detect_activities" = "angel_system.impls.detect_activities.two_stage.two_stage_detect_activities"
